AF-2: Refactoring 3
