+++ afisha-client/src/app/error/error404.component.ts
@@ -10,5 +10,4 @@ export class Error404Component implements OnInit {
   ngOnInit() {
 
   }
-
 }
