+++ afisha-client/src/app/navbar/navbar.component.html
@@ -1,10 +1,22 @@
-<mat-toolbar color="warn" role="header">
-  <span> Events Around </span>
-  <span class="example-fill-remaining-space"></span>
-  <button mat-button [matMenuTriggerFor]="animals">Members</button>
+<nav class="navbar navbar-expand-lg navbar-light bg-light">
+  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarTogglerDemo03" aria-controls="navbarTogglerDemo03" aria-expanded="false" aria-label="Toggle navigation">
+    <span class="navbar-toggler-icon"></span>
+  </button>
+  <a class="navbar-brand" href="#">Navbar</a>
 
-  <mat-menu #animals="matMenu">
-    <button mat-menu-item [routerLink] = "[ path.USER, path.LOGIN]" >Login</button>
-    <button mat-menu-item>Sign up</button>
-  </mat-menu>
-</mat-toolbar>
+  <div class="collapse navbar-collapse" id="navbarTogglerDemo03">
+    <ul class="navbar-nav mr-auto mt-2 mt-lg-0">
+      <li class="nav-item active">
+        <a class="nav-link" href="#">Home <span class="sr-only">(current)</span></a>
+      </li>
+      <li class="nav-item">
+        <a class="nav-link disabled" href="#">Disabled</a>
+      </li>
+    </ul>
+    <form class="form-inline my-2 my-lg-0">
+      <input class="form-control mr-sm-2" type="search" placeholder="Search" aria-label="Search">
+      <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
+    </form>
+    <button class="btn my-2 my-sm-0" [routerLink] = "[path.USER, path.LOGIN]">Login</button>
+  </div>
+</nav>
