+++ afisha-client/src/app/navbar/navbar.component.scss
@@ -1,5 +1,4 @@
-.example-fill-remaining-space {
-  /* This fills the remaining space, by using flexbox.
-     Every toolbar row uses a flexbox row layout. */
-  flex: 1 1 auto;
-}
+.nav.navbar-nav {font-size: 15px;}
+#searchForm {margin-right: 100px;}
+@media (max-width: 1200px) {#searchForm {display: none}}
+li > a.active {color: #F97924;}
