+++ afisha-client/src/app/shared/guards/guard-event-route.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import {ActivatedRouteSnapshot, CanActivate, Router} from "@angular/router";
-import {EventService} from "../events/services/event.service";
-import {PATH} from "./constants/path.constant";
+import {EventService} from "../../events/services/event.service";
+import {PATH} from "../constants/path.constant";
 
 @Injectable()
 export class GuardEventRouteService implements CanActivate{
