+++ afisha-client/src/app/user/login.component.ts
@@ -1,4 +1,4 @@
-import {Component, Inject, OnInit} from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from "@angular/forms";
 import {Router} from "@angular/router";
 import {PATH} from "../shared/constants/path.constant";
@@ -12,7 +12,7 @@ export class LoginComponent implements OnInit {
 
   loginForm: FormGroup;
 
-  constructor(private router: Router, @Inject(FormBuilder) formBuilder: FormBuilder) {
+  constructor(private router: Router, private formBuilder: FormBuilder) {
     this.loginForm = formBuilder.group({
       userEmail: ['', [Validators.required, Validators.email]],
       userPassword: ['', [Validators.required]]
@@ -23,26 +23,21 @@ export class LoginComponent implements OnInit {
 
   }
 
-  cancel(){
+  cancel(): void{
     this.router.navigate([PATH.EVENTS]);
   }
 
   login(loginValues) {
     if(this.validateLoginForm()){
-      //call method to login the user
-    }
-    console.log("form is not valid");
-  }
-
-  validateInputField(inputName){
-    if(inputName === 'userEmail') {
-        return this.loginForm.controls.userEmail.valid || this.loginForm.controls.userEmail.untouched;
-    } else {
-        return this.loginForm.controls.userPassword.valid || this.loginForm.controls.userPassword.untouched;
+      //@TODO: call method to login a user
     }
   }
 
-  validateLoginForm(){
+  validateInputField(inputName): boolean{
+      return this.loginForm.controls[inputName].valid || this.loginForm.controls[inputName].untouched;
+  }
+
+  validateLoginForm(): boolean{
     return this.loginForm.valid;
   }
 }
